+++ errors.go
@@ -1,0 +1,8 @@
+package main
+
+import "github.com/pkg/errors"
+
+// Custom errors.
+var (
+	ErrSignatureMismatch = errors.New("Signature mismatch")
+)
